Switch LICENSE from unlicense to MIT
The main reason I'm doing this is to avoid confusion now that I'm including some 3rd party code in the repo which is not itself released under the unlicense. To make it clearer that there's some MIT-licensed code in here too, let's just switch the whole repo. Future versions of this project will be MIT licensed.